refactor: contexts 폴더 생성
